Refactor from_scalar mod
